api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=api/requirements.txt api/requirements.in
#

annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.111.1
    # via -r api/requirements.in
fastapi-cli==0.0.4
    # via fastapi
filelock==3.15.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
flatbuffers==24.3.25
    # via onnxruntime
fsspec==2024.6.1
    # via
    #   huggingface-hub
    #   torch
gliner==0.2.8
    # via -r api/requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
huggingface-hub==0.24.2
    # via
    #   gliner
    #   tokenizers
    #   transformers
humanfriendly==10.0
    # via coloredlogs
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   fastapi
    #   torch
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.3.0
    # via python-aapy
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   onnxruntime
    #   transformers
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.5.82
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
oauthlib==3.2.2
    # via requests-oauthlib
onnxruntime==1.18.1
    # via gliner
packaging==24.1
    # via
    #   huggingface-hub
    #   onnxruntime
    #   transformers
protobuf==5.27.2
    # via onnxruntime
pydantic==2.8.2
    # via
    #   fastapi
    #   python-aapy
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via rich
python-aapy==2.0.3
    # via -r api/requirements.in
python-dotenv==1.0.1
    # via
    #   python-aapy
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.7.24
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   python-aapy
    #   requests-oauthlib
    #   transformers
requests-oauthlib==2.0.0
    # via python-aapy
rich==13.7.1
    # via typer
safetensors==0.4.3
    # via transformers
sentencepiece==0.2.0
    # via gliner
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
sympy==1.13.1
    # via
    #   onnxruntime
    #   torch
tokenizers==0.19.1
    # via transformers
torch==2.4.0
    # via gliner
tqdm==4.66.4
    # via
    #   gliner
    #   huggingface-hub
    #   python-aapy
    #   transformers
transformers==4.43.3
    # via gliner
triton==3.0.0
    # via torch
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typer
urllib3==2.2.2
    # via requests
uvicorn[standard]==0.30.3
    # via
    #   -r api/requirements.in
    #   fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
